Rename CardFinder.ts to createCard.ts
